test flight fixes

+++ ImageDataConnector/Program.cs
@@ -26,90 +26,279 @@ namespace ImageDataConnector
         static void Main(string[] args)
         {
             DatabaseHandler dbHandler = DatabaseHandler.GetInstance();
+            Console.WriteLine("Connected");
             DirectoryInfo directory = new DirectoryInfo(RECIEVING_IMAGE_FOLDER);
             FolderMonitor monitor = new FolderMonitor(directory);
-            long dataImageTimeOffset;
 
-            //wait for first image and gps data to calculate time offset
-            Console.WriteLine("Waiting for first image data for offset calculation...");
-            while(true)
+            bool canonOffsetCalculated = false;
+            long canonOffset = 0;
+            bool thermalOffsetCalculated = false;
+            long thermalOffset = 0;
+
+            List<PendingFile> canonImgList = new List<PendingFile>();
+            List<PendingFile> canonDataList = new List<PendingFile>();
+            List<PendingFile> thermalImgList = new List<PendingFile>();
+
+            
+            while (true)
             {
-                List<PendingImage> imageQueue = monitor.GetCopyOfQueue();
-                imageQueue.RemoveAll(item => item.file.Name.ToLower().Contains("thermal"));
-                if(imageQueue.Count > 0)
+                //SEPARATE NEW FILES INTO LISTS
+                List<PendingFile> fileQueue = monitor.GetCopyOfQueue();
+
+                for(int i = fileQueue.Count - 1; i >= 0; i--)
                 {
-                    int imgNum;
-                    if(!int.TryParse(imageQueue[0].file.Name.Split('_')[0], out imgNum))
+                    PendingFile pendingFile = fileQueue[i];
+                    string imgType = pendingFile.file.Name.Split('_')[0].ToLower();
+
+                    if (pendingFile.file.Extension.ToLower() == ".jpg")
                     {
-                        Console.WriteLine("Failed to parse img num: " + imageQueue[0].file.Name);
-                        continue;
+                        if (imgType == "canon")
+                            canonImgList.Add(pendingFile);
+                        else if (imgType == "thermal")
+                            thermalImgList.Add(pendingFile);
+                        else
+                            Console.WriteLine("INVALID IMG TYPE: " + imgType);
+                    }
+                    else if (pendingFile.file.Extension.ToLower() == ".imgtime")
+                    {
+                        if (imgType == "canon")
+                            canonDataList.Add(pendingFile);
+                        else
+                            Console.WriteLine("INVALID IMG TYPE: " + imgType);
                     }
 
-                    ImageData firstImageData = dbHandler.GetPhotoData(imgNum + 1);
-                    if (firstImageData != null)
-                    {
-                        Console.WriteLine("First image data recieved");
-                        DateTime cameraImageTime = ParseTimeFromImgName(imageQueue[0].file.Name);
-                        DateTime dataTime = firstImageData.DateTimeCreated;
-
-                        dataImageTimeOffset = dataTime.Ticks - cameraImageTime.Ticks + CAMERA_TRIGGER_LATENCY_MS * 10000;
-                        
-                        Console.WriteLine("Camera time: " + cameraImageTime);
-                        Console.WriteLine("Data time: " + dataTime);
-                        Console.WriteLine("Time offset (data - camera): " + dataImageTimeOffset);
-                        break;
-                    }
+                    monitor.Remove(pendingFile);
                 }
 
-                Thread.Sleep(1000);
-            }
+                //SORT LISTS
+                canonImgList.Sort((a, b) => a.file.Name.CompareTo(b.file.Name));
+                canonDataList.Sort((a, b) => a.file.Name.CompareTo(b.file.Name));
+                thermalImgList.Sort((a, b) => a.file.Name.CompareTo(b.file.Name));
 
-            int cout = 0;
-            //start main processing loop
-            while(true)
-            {
-
-                //avoid accessing the original list to avoid multithreading problems
-                List<PendingImage> imageQueue = monitor.GetCopyOfQueue();
-                Console.WriteLine("Get copy of queue " + cout++);
-
-                foreach (PendingImage image in imageQueue)
+                //CANON TIME OFFSET INITIALIZATION
+                if (!canonOffsetCalculated && canonDataList.Count > 0)
                 {
-                    Console.WriteLine("Attempting to process: " + image.file.Name);
-                    DateTime cameraTime = ParseTimeFromImgName(image.file.Name);
-                    DateTime dataTime = cameraTime.AddTicks(dataImageTimeOffset);
-
-                    Console.WriteLine("Looking up data bordering: " + dataTime.Ticks);
-                    ImageData before = dbHandler.GetClosestDataBefore(dataTime);
-                    ImageData after = dbHandler.GetClosestDataAfter(dataTime);
-
-                    if(before != null && after != null)
+                    ImageData firstTriggerData = dbHandler.GetFirstPhotoData();
+                    if (firstTriggerData != null)
                     {
-                        Console.WriteLine("Data found for: " + image.file.Name);
-                        ImageData interpolated = ImageData.interpolate(dataTime, before, after);
-
-                        //embed
-                        if( PackageAndShipImage(image, interpolated) )
+                        Console.WriteLine("First trigger data found.");
+                        long imageTime;
+                        if(GetCanonImageTime(canonDataList, canonDataList[0], out imageTime))
                         {
-                            //succesfully processed, remove from list
-                            monitor.Remove(image);
+                            long dataTime = firstTriggerData.DateTimeCreated.Ticks / 10000;
+                            canonOffset = imageTime - dataTime;
+                            canonOffsetCalculated = true;
+                            Console.WriteLine("Image time:\t" + imageTime);
+                            Console.WriteLine("Data time:\t" + dataTime);
+                            Console.WriteLine("CANON OFFSET CALCULATED: " + canonOffset);
                         }
-                        
-                    }
-                    else if (image.recieved.AddSeconds(SECS_BEFORE_SKIP_EMBED) >= DateTime.Now)
-                    {
-                        Console.WriteLine("Didn't recieve data after capture time for " + image.file.Name + " in " + SECS_BEFORE_SKIP_EMBED + " seconds. Moving to destination with bad data.");
-                        PackageAndShipImage(image, before);
                     }
                 }
 
-                //imageQueue = null;
+                //CANNON PROCESS
+                for (int i = canonImgList.Count - 1; i >= 0; i--)
+                    {
+                    PendingFile canonImg = canonImgList[i];
 
-                Thread.Sleep(1000);
+                    if(i != canonImgList.Count - 1 && canonOffsetCalculated)
+                    {
+                        long imageTime;
+                        if (GetCanonImageTime(canonDataList, canonImg, out imageTime))
+                        {
+                            DateTime time = new DateTime(imageTime * 10000 - (canonOffset*10000));
+                            ImageData before, after;
+                            if(dbHandler.GetBorderingData(time, out before, out after))
+                            {
+                                //embed
+                                ImageData.interpolate(time, before, after).SaveToImage(canonImg.file.FullName);
+                                Console.WriteLine("Embedded " + canonImg.file.Name);
+
+                                //move to embed
+                                try
+                                {
+                                    File.Move(canonImg.file.FullName, DESTINATION_IMAGE_FOLDER + canonImg.file.Name);
+                                }
+                                catch (Exception e)
+                                {
+                                    Console.WriteLine("Failed to move embedded image to image folder " + e);
+                                }
+
+                                //copy to embed backup
+                                try
+                                {
+                                    if (!File.Exists(DESTINATION_IMAGE_FOLDER_BACKUP + canonImg.file.Name))
+                                        File.Copy(DESTINATION_IMAGE_FOLDER + canonImg.file.Name, DESTINATION_IMAGE_FOLDER_BACKUP + canonImg.file.Name);
+                                }
+                                catch (Exception e)
+                                {
+                                    Console.WriteLine("Failed to backup embedded image to backup embedded folder " + e);
+                                }
+
+                                canonImgList.RemoveAt(i);
+                                continue;
+                            }
+                        }
+                    }
+
+                    //if an image has been without data for too long, send it with no embedding
+                    if (DateTime.Now >= canonImg.recieved.AddSeconds(SECS_BEFORE_SKIP_EMBED))
+                    {
+                        Console.WriteLine(canonImg.file.Name + " reached data timeout"); //could still have valid data, this just means the data for the next image wasn't received (which ensures this data is valid)
+                        long imageTime;
+                        if (GetCanonImageTime(canonDataList, canonImg, out imageTime))
+                        {
+                            DateTime time = new DateTime(imageTime * 10000 - canonOffset*10000);
+                            ImageData before, after;
+                            if (dbHandler.GetBorderingData(time, out before, out after))
+                            {
+                                //embed
+                                ImageData.interpolate(time, before, after).SaveToImage(canonImg.file.FullName);
+                                Console.WriteLine("Embedded " + canonImg.file.Name);
+                            }
+                        }
+
+                        //move to embed
+                        try
+                        {
+                            File.Move(canonImg.file.FullName, DESTINATION_IMAGE_FOLDER + canonImg.file.Name);
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine("Failed to move embedded image to image folder " + e);
+                        }
+
+                        //copy to embed backup
+                        try
+                        {
+                            if (!File.Exists(DESTINATION_IMAGE_FOLDER_BACKUP + canonImg.file.Name))
+                                File.Copy(DESTINATION_IMAGE_FOLDER + canonImg.file.Name, DESTINATION_IMAGE_FOLDER_BACKUP + canonImg.file.Name);
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine("Failed to backup embedded image to backup embedded folder " + e);
+                        }
+
+                        canonImgList.RemoveAt(i);
+                    }
+                }
+
+                //THERMAL TIME OFFSET INITIALIZATION
+                if (!thermalOffsetCalculated && thermalImgList.Count > 0)
+                {
+                    ImageData firstTriggerData = dbHandler.GetFirstPhotoData();
+                    if (firstTriggerData != null)
+                    {
+                        Console.WriteLine("First trigger data found.");
+                        long imageTime;
+                        if (Int64.TryParse(thermalImgList[0].file.Name.Split('_')[1].Replace(thermalImgList[0].file.Extension, ""), out imageTime))
+                        {
+                            long dataTime = firstTriggerData.DateTimeCreated.Ticks / 10000;
+                            thermalOffset = imageTime - dataTime;
+                            thermalOffsetCalculated = true;
+                            Console.WriteLine("Image time:\t" + imageTime);
+                            Console.WriteLine("Data time:\t" + dataTime);
+                            Console.WriteLine("THERMAL OFFSET CALCULATED: " + thermalOffset);
+                        }
+                    }
+                }
+
+                //THERMAL PROCESSING
+                for (int i = thermalImgList.Count - 1; i >= 0; i--)
+                {
+                    PendingFile thermalImg = thermalImgList[i];
+
+                    if (thermalOffsetCalculated)
+                    {
+                        long imageTime;
+                        if (Int64.TryParse(thermalImgList[0].file.Name.Split('_')[1] ,out imageTime))
+                        {
+                            DateTime time = new DateTime(imageTime * 10000 - canonOffset*10000);
+                            ImageData before, after;
+                            if (dbHandler.GetBorderingData(time, out before, out after))
+                            {
+                                //embed
+                                ImageData.interpolate(time, before, after).SaveToImage(thermalImg.file.FullName);
+                                Console.WriteLine("Embedded " + thermalImg.file.Name);
+
+                                //move to embed
+                                try
+                                {
+                                    File.Move(thermalImg.file.FullName, DESTINATION_IMAGE_FOLDER + thermalImg.file.Name);
+                                }
+                                catch (Exception e)
+                                {
+                                    Console.WriteLine("Failed to move embedded image to image folder " + e);
+                                }
+
+                                //copy to embed backup
+                                try
+                                {
+                                    File.Copy(DESTINATION_IMAGE_FOLDER + thermalImg.file.Name, DESTINATION_IMAGE_FOLDER_BACKUP + thermalImg.file.Name);
+                                }
+                                catch (Exception e)
+                                {
+                                    Console.WriteLine("Failed to backup embedded image to backup embedded folder " + e);
+                                }
+
+                                thermalImgList.RemoveAt(i);
+                                continue;
+                            }
+                        }
+                    }
+
+                    //if an image has been without data for too long, send it with no embedding
+                    if (DateTime.Now >= thermalImg.recieved.AddSeconds(SECS_BEFORE_SKIP_EMBED))
+                    {
+                        Console.WriteLine(thermalImg.file.Name + " reached data timeout");
+                        //move to embed
+                        try
+                        {
+                            File.Move(thermalImg.file.FullName, DESTINATION_IMAGE_FOLDER + thermalImg.file.Name);
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine("Failed to move embedded image to image folder " + e);
+                        }
+
+                        //copy to embed backup
+                        try
+                        {
+                            if (!File.Exists(DESTINATION_IMAGE_FOLDER_BACKUP + thermalImg.file.Name))
+                                File.Copy(DESTINATION_IMAGE_FOLDER + thermalImg.file.Name, DESTINATION_IMAGE_FOLDER_BACKUP + thermalImg.file.Name);
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine("Failed to backup embedded image to backup embedded folder " + e);
+                        }
+
+                        thermalImgList.RemoveAt(i);
+                    }
+                }
             }
         }
 
-        private static bool PackageAndShipImage(PendingImage image, ImageData data)
+        private static bool GetCanonImageTime(List<PendingFile> canonDataList, PendingFile image, out long imageTime)
+        {
+            imageTime = 0;
+            for (int i = canonDataList.Count - 1; i >= 0; i--)
+            {
+                PendingFile pendingFile = canonDataList[i];
+                if (pendingFile.file.Name.Split('_')[1].ToLower() == image.file.Name.Split('_')[1].Replace(image.file.Extension, ""))
+                {
+                    if (Int64.TryParse(pendingFile.file.Name.Split('_')[2].Replace(pendingFile.file.Extension, ""), out imageTime))
+                    {
+                        return true;
+                    }
+
+                    Console.WriteLine("FAILED TO PARSE TIME FROM: " + pendingFile.file.Name);
+                    return false;
+                }
+            }
+
+            return false;
+        }
+
+        private static bool PackageAndShipImage(PendingFile image, ImageData data)
         {
             //embed
             if (data != null)
@@ -139,7 +328,10 @@ namespace ImageDataConnector
             Console.WriteLine("Moving " + image.file.Name + " to " + destinationFile);
             try
             {
-                System.IO.File.Move(image.file.FullName, destinationFile);
+                if (!System.IO.File.Exists(destinationFile))
+                    System.IO.File.Move(image.file.FullName, destinationFile);
+
+                if (!System.IO.File.Exists(DESTINATION_IMAGE_FOLDER_BACKUP + image.file.Name))
                 System.IO.File.Copy(destinationFile, DESTINATION_IMAGE_FOLDER_BACKUP + image.file.Name);
             }
             catch (Exception e)
